+++ spec/system/example_wondrous_forest_spec.rb
@@ -55,8 +55,8 @@ ceedling_system_tests do
         @c.with_context do
           Dir.chdir "wondrous_forest" do
             @output = @c.ceedling_build_exec("test:all")
-            expect(@output).to match(/TESTED:\s+65/)
-            expect(@output).to match(/PASSED:\s+65/)
+            expect(@output).to match(/TESTED:\s+67/)
+            expect(@output).to match(/PASSED:\s+67/)
             expect(@output).to match(/FAILED:\s+0/)
           end
         end